% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_units.R
\name{add_units}
\alias{add_units}
\title{Add Unit Mosaics to an Existing List of Units}
\usage{
add_units(list_units_old, supercells, range_n_toadd)
}
\arguments{
\item{list_units_old}{A list of existing unit mosaics, each being a data.frame or similar object.}

\item{supercells}{An object containing the supercell structure used to generate new unit mosaics.}

\item{range_n_toadd}{A numeric vector indicating the sizes of new units to add from the supercells.}
}
\value{
A list with two elements:
\describe{
\item{range_n}{A numeric vector containing the sizes (number of rows) of all unit mosaics (old and new).}
\item{list_units}{A list combining the old and new unit mosaics.}
}
}
\description{
This function extends an existing list of unit mosaics by adding new units
generated from supercells based on a specified range of unit sizes.
}
\examples{
# Assuming you have defined `supercells`, `list_units_old`, and `range_n_toadd`:
# result <- add_units(list_units_old, supercells, range_n_toadd)

}
