% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ms.R
\name{compute_ms}
\alias{compute_ms}
\title{Compute the meso-size (ms) index}
\usage{
compute_ms(c_logmean, c_logsd, W_area, A_tot, n_range, n_iter = 500)
}
\arguments{
\item{c_logmean}{Mean of the log-normal distribution (on log scale).}

\item{c_logsd}{Standard deviation of the log-normal distribution (on log scale).}

\item{W_area}{reference area (width^2)}

\item{A_tot}{Numeric value, total reference (reach) area.}

\item{n_range}{Integer vector of units range}

\item{n_iter}{Number of iterations (default = 500).}
}
\value{
A tibble with columns:
\describe{
\item{n}{n units}
\item{ms}{ms index}
}
}
\description{
This function simulates areas based on log-normal random sampling (empirical meso-sizes c)
and computes a normalized index of deviation from a target area (total reach area)
}
\details{
For each unit number in \code{n_range}, the function simulates random draws from
the empirically defined log-normal distribution, computes probable resulting area,
and compares the resulting area to the reference total \code{A_tot}.
}
