% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_widths.R
\name{compute_widths}
\alias{compute_widths}
\title{Compute Average Channel Width from Wetted Channel Polygon}
\usage{
compute_widths(
  poly_wetchannel,
  centerline,
  trans_longdist = 10,
  trans_width = 300
)
}
\arguments{
\item{poly_wetchannel}{An \code{sf} POLYGON object representing the wetted channel area.}

\item{centerline}{An \code{sf} LINESTRING object representing the channel centerline.}

\item{trans_longdist}{Numeric. Longitudinal spacing between transects (default is 10, in CRS units).}

\item{trans_width}{Numeric. Width of transects to extend on either side of the centerline
(default is 300, in CRS units).}
}
\value{
Numeric value representing the \strong{average channel width} based on the intersections of transects.
}
\description{
This function calculates the average channel width by intersecting evenly spaced
transects with a polygon representing the wetted channel.
}
\details{
\itemize{
\item The function uses \code{centerline_to_transects()} to generate perpendicular cross-sections.
\item These transects are intersected with the wetted channel polygon to extract segments.
\item The average length of these intersected segments is returned as the representative channel width.
}
}
\examples{
# Example usage:
# avg_width <- compute_widths(poly_wetchannel, centerline, trans_longdist = 10, trans_width = 300)

}
