% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nmosaic.R
\name{create_nmosaic}
\alias{create_nmosaic}
\title{Extract mosaics from SKATER cut matrix}
\usage{
create_nmosaic(supercells, res_allcuts, n_range)
}
\arguments{
\item{supercells}{spatial object containing supercells}

\item{res_allcuts}{output object from \code{skater_allcuts}, including
\code{groups.history}}

\item{n_range}{integer vector defining n values}
}
\value{
List of segmented mosaics with n units defined in \code{n_range}.
}
\description{
Extracts spatial mosaic partitions with n units defined in \code{n_range}
}
