% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_holes.R
\name{fill_holes}
\alias{fill_holes}
\title{Fill Holes in Habitat Polygons}
\usage{
fill_holes(poly_mesohabitats, file_name, threshold = 10000)
}
\arguments{
\item{poly_mesohabitats}{An \code{sf} object containing polygon geometries
representing mesohabitats.}

\item{file_name}{Character string giving the output shapefile path.}

\item{threshold}{Numeric value giving the maximum area (in map units) of holes
to fill. Defaults to 10,000 (map units)}
}
\value{
Invisibly returns the modified \code{sf} object (without holes).
}
\description{
This function removes interior holes from polygon features so that
they can be processed by SimStream-Web, which do not handle
polygons with holes correctly. The function also writes the resulting
polygon layer to a shapefile.
}
\details{
Internally, this function uses \code{smoothr::fill_holes()} to remove holes
smaller than the specified threshold. The modified polygon is then written
to disk using \code{sf::st_write()} with \code{delete_layer = TRUE}.
}
\examples{
\dontrun{
  filled <- fill_holes(poly_mesohabitats, "mesohabitats_noholes.shp", threshold = 5000)
}

}
