% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_n.R
\name{optimal_n}
\alias{optimal_n}
\title{Identify Optimal 'n' Value Based on Global Score (GS)}
\usage{
optimal_n(optimal_metrics)
}
\arguments{
\item{optimal_metrics}{A list or data structure containing:
\describe{
\item{GS}{A numeric vector of GS values.}
\item{n}{A vector of associated parameter values for \code{n}.}
}}
}
\value{
A data frame containing:
\describe{
\item{i}{The index of the optimal \code{n} value.}
\item{n}{The optimal value of \code{n} that minimizes \code{GS}.}
}
}
\description{
Extracts the optimal value of the parameter \code{n} (number of units) corresponding to the minimum value
of the Global Score
}
\details{
The function identifies the index of the minimum \code{GS} value and returns the corresponding \code{n} value at that index.
}
\examples{
gs_values <- c(0.3, 0.2, 0.15, 0.25)
n_values <- seq(1, 4)
optimal_metrics <- data.frame(GS = gs_values, n = n_values)
optimal_n(optimal_metrics)

}
