% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell_to_units.R
\name{supercell_to_units}
\alias{supercell_to_units}
\title{Regionalize supercells into mosaic of n units Using SKATER Algorithm}
\usage{
supercell_to_units(mesh_supercell, range_n)
}
\arguments{
\item{mesh_supercell}{A \code{sf} object containing polygons (supercells) with at least columns \code{D_avg} and \code{V_avg}
representing average depth and velocity, respectively.}

\item{range_n}{A numeric vector specifying the target number of regions for each SKATER run.}
}
\value{
A list of \code{sf} objects, each representing a segmentation into \code{n} spatial units (as defined in \code{range_n}),
with recalculated average depth (\code{D_avg}) and velocity (\code{V_avg}) for each unit.
}
\description{
Applies spatial regionalization to supercell polygons using the SKATER algorithm.
This function segments a mesh of supercells into spatial units based on
average depth and velocity attributes.
}
\details{
The function performs the following steps:
\enumerate{
\item Selects relevant variables (\code{D_avg}, \code{V_avg}).
\item Builds neighbor relationships using \code{poly2nb()}.
\item Scales input variables and computes edge costs.
\item Constructs a minimum spanning tree.
\item Applies the SKATER algorithm for each target number of regions.
}
}
