% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect.R
\name{transect}
\alias{transect}
\title{Compute Transects Perpendicular to a Line}
\usage{
transect(tpts, tlen)
}
\arguments{
\item{tpts}{A \code{data.frame} or \code{tibble} containing the coordinates (\code{x}, \code{y}) and orientation angle (\code{theta}) of each point.}

\item{tlen}{Numeric. Half-length of the transects to be generated (i.e., total transect length will be \code{2 * tlen}).}
}
\value{
A \code{data.frame} with columns \code{x0}, \code{y0}, \code{x1}, \code{y1}, representing the start and end coordinates of each transect line.
}
\description{
Generates transects perpendicular to a given point with associated angle directions.
Each transect is centered at the input point and extends in both directions by the specified length.
}
\details{
The function calculates a perpendicular angle to \code{theta} (\code{thetaT = theta + π/2}) and offsets the coordinates accordingly to create a line segment.
}
