% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_optimality_metrics.R
\name{compute_optimality_metrics}
\alias{compute_optimality_metrics}
\title{Compute Normalized Optimality Metrics (v,MI,ms) and combine into Global Score}
\usage{
compute_optimality_metrics(flow, flow_regions, range_n, c_params)
}
\arguments{
\item{flow}{An \code{sf} object (sfc POLYGON) representing modeled flow areas (e.g., hydraulic model elements).}

\item{flow_regions}{A list of mosaic region groupings (output from \code{supercell_to_units()}).}

\item{range_n}{A numeric vector of unit counts associated with each mosaic.}

\item{c_params}{A data frame containing the parameters W_area, c_logmean and c_logsd}
}
\value{
A data frame containing:
\describe{
\item{n}{Number of units per mosaic.}
\item{MI}{Normalized Moran's I}
\item{v}{Normalized weighted averaged variance}
\item{ms}{Normalized meso-size index}
\item{GS}{Global Score, calculated as the mean of MI, v, and ms.}
}
}
\description{
This function calculates three normalized optimality metrics, the Moran's I (MI), weighted averaged variance (v), and meso-size index (ms), for a list of mosaic regions and combines
them into a Global Score (GS).
}
\details{
\itemize{
\item MI and v are computed using the \code{segmentation_indicators()} function on each mosaic.
\item ms is calculated based on an expected log-normal distribution.
\item All three metrics are normalized using \code{sp_norm()} and combined into a global performance score (GS).
}
}
\examples{
# Assuming 'flow', 'flow_regions', 'range_n', and other parameters are defined:
# scores <- compute_optimality_metrics(flow, flow_regions, range_n, W_area, c_logmean, c_logsd)

}
