% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_DVS.R
\name{extract_DVS}
\alias{extract_DVS}
\title{Prepare Depth, Velocity, Substrate (DVS) Table for MesoHABSIM}
\usage{
extract_DVS(
  poly_mesohabitats,
  flow,
  ras_substrate,
  sub_codes,
  file_name = "mesohabitats_DVS.txt",
  nmax = 50
)
}
\arguments{
\item{poly_mesohabitats}{sf object of polygons representing mesohabitats}

\item{flow}{sf object of mesh elements with flow data (DEPTH, VEL)}

\item{ras_substrate}{stars object of substrate raster}

\item{sub_codes}{data.frame with columns \code{class_id} and \code{type} for substrate classification}

\item{file_name}{character string, path and file name to save output table. Default: 'mesohabitats_DVS.txt'}

\item{nmax}{numeric, maximum number of points per HMU for subsampling (default 50)}
}
\description{
Prepare Depth, Velocity, Substrate (DVS) Table for MesoHABSIM
}
